experimental/live-lab/training/requirements.txt
joblib>=1.4
numpy>=2.0
Pillow>=10.0
scikit-learn>=1.5
